refector: 향수 리뷰작성 리팩토링
